Update allowed origins and database path
This commit introduces several changes necessary for deploying the app locally. In api/app.py, localhost addresses have been added to the 'origins' list to enable CORS for local development.
The database path was also updated to reflect the correct location. Besides, in Dockerfile, the exposed port number is changed to 8037 to match the usual dev port and the unnecessary author label and topic_models copy instruction are removed for cleaner code.